+++ src/redisenterprise/azext_redisenterprise/aaz/latest/redisenterprise/database/_export.py
@@ -19,9 +19,9 @@ class Export(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-05-01-preview",
+        "version": "2025-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cache/redisenterprise/{}/databases/{}/export", "2025-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cache/redisenterprise/{}/databases/{}/export", "2025-07-01"],
         ]
     }
 
@@ -157,7 +157,7 @@ class Export(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-05-01-preview",
+                    "api-version", "2025-07-01",
                     required=True,
                 ),
             }